+++ BankingApp/BankingAppApiService/StaffApiService/src/main/java/com/learning/service/StaffService.java
@@ -1,15 +1,13 @@
 package com.learning.service;
 
 import com.learning.Entity.model.StaffEntity;
-
 import java.util.List;
-import java.util.Optional;
 
 public interface StaffService {
 
-    public List<StaffEntity> viewStaff(Integer approverId);
+    public List<StaffEntity> viewStaff(Long approverId);
 
     public void createStaff(StaffEntity staff);
 
-    public void changeStatus(Integer staffId, boolean status);
+    public void changeStatus(Long staffId, boolean status);
 }
